Move ShutterControlState into dto subfolder and remove @NonNullByDefault annotation
